libft norminette

+++ libft/ft_strnlen.c
@@ -12,9 +12,9 @@
 
 #include "libft.h"
 
-size_t		ft_strnlen(const char *str, size_t maxlen)
+size_t	ft_strnlen(const char *str, size_t maxlen)
 {
-	unsigned len;
+	size_t	len;
 
 	len = 0;
 	while (len < maxlen && *str)
